Moved Results.js component files
